refactor(ci): move from sleep to kafka lag based testing (#8094)
