--- smoke-test/requests_wrapper/constants.py
@@ -1,2 +1,0 @@
-
-ELASTICSEARCH_REFRESH_INTERVAL_SECONDS = 4
