+++ smoke-test/tests/cli/delete_cmd/test_timeseries_delete.py
@@ -14,7 +14,7 @@ from datahub.entrypoints import datahub
 from datahub.metadata.schema_classes import DatasetProfileClass
 from tests.aspect_generators.timeseries.dataset_profile_gen import \
     gen_dataset_profiles
-from tests.utils import get_strftime_from_timestamp_millis
+from tests.utils import get_strftime_from_timestamp_millis, wait_for_writes_to_sync
 import requests_wrapper as requests
 
 logger = logging.getLogger(__name__)
@@ -31,8 +31,7 @@ runner = CliRunner(mix_stderr=False)
 
 
 def sync_elastic() -> None:
-    time.sleep(requests.ELASTICSEARCH_REFRESH_INTERVAL_SECONDS)
-
+    wait_for_writes_to_sync()
 
 def datahub_put_profile(dataset_profile: DatasetProfileClass) -> None:
     with tempfile.NamedTemporaryFile("w+t", suffix=".json") as aspect_file:
