+++ smoke-test/tests/cli/user_groups_cmd/test_group_cmd.py
@@ -10,12 +10,13 @@ from datahub.entrypoints import datahub
 from datahub.ingestion.graph.client import DataHubGraph, get_default_graph
 import time
 import requests_wrapper as requests
+from tests.utils import wait_for_writes_to_sync
 
 runner = CliRunner(mix_stderr=False)
 
 
 def sync_elastic() -> None:
-    time.sleep(requests.ELASTICSEARCH_REFRESH_INTERVAL_SECONDS)
+    wait_for_writes_to_sync()
 
 
 def datahub_upsert_group(group: CorpGroup) -> None:
